Test/ES6: separate some tests for not ES6 comptability (READ COMMIT)
When I was writing tests for TornadoWebSocket.Module, I tested the inheritance of this class. 
But I encountered a problem of comptability: for the non ES5, I should use the old method "MyModule = function(TornadoWebSocket.Module){ inherits(MyModule, TornadoWebSocket.Module); function MyModule(prefix) { ... } 
}" with a Babel's method to inherit MyModule with TornadoWebSocket.Module.
It works very fine, but when I was testing MyModule during ES6 tests, I got an error message telling me that I should use the `new` keyword  (during TornadoWebSocket.Module instanciation).
So I tried to use the env' variable called `USE_ES6`, and use the old way when `USE_ES6` is off,  and use `class MyModule extends TornadoWebSocket.Module` when it is on.
But, it didn't works on Travis. When my tests ran on Travis during non ES6 tests (so « old » browsers), the keyword `class` was reserved, and tests failed.

So, I take a decision to implement a ES6 and non ES6 version of some tests (here for tws.Module) in separate files, and use the file `test-main.js` to dynamically load them. 
It works fine, but tests are duplicated, so I will try to find a better and cleaner way to resolve this problem.